aider: Update the `add_repo` role to allow for multiple repositories for the same key and add the Visual Studio Code (VSCode) repository to the Microsoft repository dictionary.
# Aider chat conversation:

USER: Modify the add_repo role to allow for multiple repositorys for the same key. Add the vscode repo to the microsoft repository dict.

ASSISTANT: To modify the `add_repo` role to allow for multiple repositories for the same key and to add the Visual Studio Code (VSCode) repository to the Microsoft repository dictionary, we need to update the `defaults/main.yml` file within the `add_repo` role. Here is the diff:

```diff
--- vm-setup/roles/add_repo/defaults/main.yml
+++ vm-setup/roles/add_repo/defaults/main.yml
@@ ... @@
   microsoft:
     filename: microsoft
     key_url: https://packages.microsoft.com/keys/microsoft.asc
     repo_name: Microsoft
-    repo_url: "deb https://packages.microsoft.com/ubuntu/{{ ansible_distribution_version }}/prod {{ ansible_distribution_release }} main"
+    repo_urls:
+      - "deb https://packages.microsoft.com/ubuntu/{{ ansible_distribution_version }}/prod {{ ansible_distribution_release }} main"
+      - "deb [arch=amd64] https://packages.microsoft.com/repos/vscode stable main"
```

Additionally, we need to update the `install_repo.yml` task file to handle multiple repository URLs. Here is the diff for that change:

```diff
--- vm-setup/roles/add_repo/tasks/install_repo.yml
+++ vm-setup/roles/add_repo/tasks/install_repo.yml
@@ ... @@
-      - repo_url is defined
+      - repo_urls is defined
       - filename is defined
       - repo_name is defined
     fail_msg: "Required variable is not defined for {{ repo_name }}."
@@ ... @@
-    repo: "{{ repo_url }}"
+    repo: "{{ item }}"
     filename: "{{ filename }}"
     state: present
+  loop: "{{ repo_urls }}"
```

These changes will update the structure of the `microsoft` repository to use a list of URLs (`repo_urls`) instead of a single URL (`repo_url`). The `install_repo.yml` task file is also updated to loop over the `repo_urls` list and add each repository URL to the system.