aider: Added the Docker repository to the add_repo role.
# Aider chat conversation:

USER: Add the docker repository to the add_repo role.

ASSISTANT: To add the Docker repository to the `add_repo` role, we need to update the `defaults/main.yml` file within the `add_repo` role to include the Docker repository details.

Here is the diff for this change:

```diff
--- vm-setup/roles/add_repo/defaults/main.yml
+++ vm-setup/roles/add_repo/defaults/main.yml
@@ ... @@
   microsoft:
     filename: microsoft
     key_url: https://packages.microsoft.com/keys/microsoft.asc
     repo_name: Microsoft
     repo_urls:
       - "deb https://packages.microsoft.com/ubuntu/{{ ansible_distribution_version }}/prod {{ ansible_distribution_release }} main"
       - "deb https://packages.microsoft.com/repos/vscode stable main"
+  docker:
+    filename: docker
+    key_url: https://download.docker.com/linux/ubuntu/gpg
+    repo_name: Docker
+    repo_urls:
+      - "deb [arch=amd64] https://download.docker.com/linux/ubuntu {{ ansible_distribution_release }} stable"
```

This change adds a new entry for the Docker repository in the `repos` dictionary, specifying the filename for the list file, the URL for the GPG key, the human-readable name of the repository, and the repository URLs.