aider: Added task to the `gcloud` role to install the `gcloud` group using the `install_pkgs` role.
# Aider chat conversation:

USER: Modify the gcloud role to use the install_pkgs role to install the gcloud group

ASSISTANT: To modify the `gcloud` role to use the `install_pkgs` role to install the `gcloud` group, we need to update the `vm-setup/roles/gcloud/tasks/main.yml` file. Here is the diff for the required changes:

```diff
--- vm-setup/roles/gcloud/tasks/main.yml
+++ vm-setup/roles/gcloud/tasks/main.yml
@@ ... @@
 # This playbook uses the add_repo role to install the Google Cloud repository
 - name: Include add_repo to install Google Cloud repository
   include_role:
     name: add_repo
   vars:
     repo_list:
       - google-cloud
+
+- name: Install Google Cloud SDK packages
+  include_role:
+    name: install_pkgs
+  vars:
+    package_groups: "gcloud"
```

This change adds a task to the `gcloud` role that includes the `install_pkgs` role with the `package_groups` variable set to `"gcloud"`, which corresponds to the group we previously defined in the `install_pkgs` role.