aider: Add call to autostart task in gnome settings role to autostart Tilix.
# Aider chat conversation:

USER: Add an call to the autostart task in the gnome settings role to autostart tilix.

ASSISTANT: To add a call to the autostart task in the `gnome_settings` role to autostart Tilix, I will modify the `main.yml` file within the `tasks` directory of the `gnome_settings` role. The task will include the `create_autostart.yml` task file and provide the necessary variables to create an autostart entry for Tilix.

Here is the diff for the changes:

```diff
--- vm-setup/roles/gnome_settings/tasks/main.yml
+++ vm-setup/roles/gnome_settings/tasks/main.yml
@@ ... @@

 - name: Set titlebar font
   command: gsettings set org.gnome.desktop.wm.preferences titlebar-uses-system-font {{ gnome_settings_titlebar_uses_system_font | lower }}

+- name: Autostart Tilix on login
+  include_tasks: create_autostart.yml
+  vars:
+    autostart_app_name: "Tilix"
+    autostart_app_exec: "tilix"
+    autostart_app_comment: "Tilix terminal emulator"
+    autostart_app_terminal: false
+    autostart_app_hidden: false
+
 - name: Disable services
   become: yes
   systemd:
     name: "{{ item }}"
     enabled: no
     masked: yes
     state: stopped
   loop: "{{ gnome_settings_disabled_services }}"
```

This change will ensure that when the `gnome_settings` role is run, it will also create an autostart entry for Tilix, causing it to start automatically when the user logs into their GNOME session.