aider: Added dock settings to gnome settings to enable auto-hide, set dock icon size to 40, position dock to the left, and disable show volumes and devices.
# Aider chat conversation:

USER: Add dock settings to gnome settings that enable auto-hide the dock and panel mod, sets dock icon size to 40, sets the position to 'left' and disables show volumes and devices.

ASSISTANT: To add the dock settings to the `gnome_settings` role, I will modify the `main.yml` file within the `tasks` directory of the `gnome_settings` role. The changes will include setting the auto-hide feature for the dock, setting the dock icon size, positioning the dock to the left, and disabling the display of volumes and devices on the dock.

Here is the diff for the changes:

```diff
--- vm-setup/roles/gnome_settings/tasks/main.yml
+++ vm-setup/roles/gnome_settings/tasks/main.yml
@@ ... @@
 - name: Set titlebar font
   command: gsettings set org.gnome.desktop.wm.preferences titlebar-uses-system-font {{ gnome_settings_titlebar_uses_system_font | lower }}

+- name: Enable auto-hide for the dock
+  command: gsettings set org.gnome.shell.extensions.dash-to-dock autohide true
+
+- name: Enable panel mode for the dock
+  command: gsettings set org.gnome.shell.extensions.dash-to-dock extend-height false
+
+- name: Set dock icon size
+  command: gsettings set org.gnome.shell.extensions.dash-to-dock dash-max-icon-size 40
+
+- name: Set dock position to left
+  command: gsettings set org.gnome.shell.extensions.dash-to-dock dock-position 'LEFT'
+
+- name: Disable show volumes and devices on the dock
+  command: gsettings set org.gnome.shell.extensions.dash-to-dock show-mounts false
+
 - name: Disable services
   become: yes
   systemd:
```

These changes will configure the GNOME dock according to the specified settings when the `gnome_settings` role is executed.