aider: Update default variables to match looping structure of desktop interface settings.
# Aider chat conversation:

USER: Modify the default variables to be the same looping nature of desktop interface.

ASSISTANT: To modify the default variables in the `gnome_settings` role to use a looping structure similar to the desktop interface settings, I will update the `vm-setup/roles/gnome_settings/defaults/main.yml` file. Here's the diff for the changes:

```diff
--- vm-setup/roles/gnome_settings/defaults/main.yml
+++ vm-setup/roles/gnome_settings/defaults/main.yml
@@ ... @@
-gnome_settings_dock_autohide: true
-gnome_settings_dock_extend_height: false
-gnome_settings_dock_icon_size: 40
-gnome_settings_dock_position: 'LEFT'
-gnome_settings_dock_show_mounts: false
+gnome_settings_dock:
+  - { key: 'autohide', value: true }
+  - { key: 'extend-height', value: false }
+  - { key: 'dash-max-icon-size', value: 40 }
+  - { key: 'dock-position', value: 'LEFT' }
+  - { key: 'show-mounts', value: false }
```

This diff replaces the individual variables for each dash-to-dock setting with a list of dictionaries, each containing a `key` and a `value`. This structure is consistent with the looping nature of the desktop interface settings in the same file.