aider: Moved the `gcloud` install group to the `gcloud` role to call `install_pkgs` with the list of files to be installed.
# Aider chat conversation:

USER: Move the gcloud install group to the gcloud role such that it calls install_pks with the list of files to be installed.

ASSISTANT: To move the `gcloud` install group to the `gcloud` role such that it calls `install_pkgs` with the list of files to be installed, we need to make changes to the `gcloud` role's `tasks/main.yml` file. We will remove the direct variable assignment for `package_groups` and instead pass the list of packages as a parameter to the `install_pkgs` role.

Here are the diffs for the required changes:

```diff
--- vm-setup/roles/gcloud/tasks/main.yml
+++ vm-setup/roles/gcloud/tasks/main.yml
@@ ... @@
 # This playbook uses the add_repo role to install the Google Cloud repository
 - name: Include add_repo to install Google Cloud repository
   include_role:
     name: add_repo
   vars:
     repo_list:
       - google-cloud

-- name: Install Google Cloud SDK packages
-  include_role:
-    name: install_pkgs
-  vars:
-    package_groups: "gcloud"
+- name: Define Google Cloud SDK packages
+  set_fact:
+    gcloud_packages:
+      - google-cloud-cli
+
+- name: Install Google Cloud SDK packages using install_pkgs role
+  include_role:
+    name: install_pkgs
+    apply:
+      vars:
+        package_groups: "{{ gcloud_packages }}"
```

This change does the following:

1. Defines a new variable `gcloud_packages` with the list of Google Cloud SDK packages to be installed.
2. Modifies the `include_role` task to use the `install_pkgs` role, passing the `gcloud_packages` list as a variable.

With these changes, the `gcloud` role will now call the `install_pkgs` role with the specific list of packages to install, which aligns with the desired behavior.