aider: Add apt module to install requested packages in install_pkgs role.
# Aider chat conversation:

USER: I ran this command:

 cd vm-setup && ansible-playbook -K test-role.yml

And got this output:

ERROR! no module/action detected in task.

The error appears to be in '/home/harleypig/projects/dotfiles/vm-setup/roles/install_pkgs/tasks/main.yml': line 2, column 3, but may
be elsewhere in the file depending on the exact syntax problem.

The offending line appears to be:

---
- name: Install packages
  ^ here

ASSISTANT: The error message indicates that there is a task in the `install_pkgs` role's `main.yml` file that does not have an associated module or action. This typically means that the task is missing a module such as `apt`, `yum`, or any other that should be performing an action.

To resolve this, we need to add an appropriate module to the task. Since the `install_pkgs` role seems to be designed for installing packages, I will add the `apt` module to the task with a placeholder for the package names. If the package names are meant to be provided by a variable, I will use the `packages` variable as it is commonly used for such purposes.

Here is the diff to fix the issue:

```diff
--- vm-setup/roles/install_pkgs/tasks/main.yml
+++ vm-setup/roles/install_pkgs/tasks/main.yml
@@ -1,4 +1,5 @@
 ---
-- name: Install packages
+- name: Install requested packages
+  apt:
+    state: latest
+    update_cache: yes
+    name: "{{ packages }}"
```

This change assumes that the `packages` variable will be provided with a list of package names to install. If the variable name is different or if there are other requirements, please let me know so I can adjust the diff accordingly.