aider: Updated the dependency in gnome_settings to pass the packages variable correctly to install_pkgs.
# Aider chat conversation:

USER: All you did was change the variable name. I'm getting the same error, just with the new variable name.

ASSISTANT: The error you're encountering suggests that the `gnome_settings` role is not passing the `packages` variable to the `install_pkgs` role correctly. To ensure that the `packages` variable is set and passed correctly, we need to look at how the `gnome_settings` role is including the `install_pkgs` role and how the variable is defined.

From the provided information, it seems that the `gnome_settings` role has a `meta/main.yml` file that defines the dependency on the `install_pkgs` role and attempts to pass the `packages` variable. However, since you're still encountering the error, there might be an issue with the way the variable is being passed or defined.

To troubleshoot this, we need to ensure that the `gnome_settings` role's `defaults/main.yml` or `vars/main.yml` (if it exists) defines the `packages` variable or that it's being passed in from a playbook. Since the `gnome_settings` role's `defaults/pkgs.yml` defines a `gnome_package_group` variable, we need to make sure that this variable is being correctly mapped to `packages` when the `install_pkgs` role is included.

Please provide the `gnome_settings` role's `meta/main.yml` and any playbook that includes the `gnome_settings` role so that I can propose the correct changes to resolve the issue. If the `gnome_settings` role's `meta/main.yml` is not read-write, please add it to the chat so I can make the necessary edits.

USER: The install packages role is supposed to accept a list of packages to install. It works ... I've tested it from a playbook. I want other roles to define the list of packages to install in a separate defaults file (as gnome settings is doing now). Does the call to install packages have to come from the meta directory? Don't make any changes.

ASSISTANT: No, the call to the `install_pkgs` role does not have to come from the `meta` directory. In Ansible, the `meta/main.yml` within a role is typically used to define role dependencies, which are roles that should automatically be applied before the role itself is applied. However, this is not the only way to include roles.

Roles can also be included directly in playbooks using the `include_role` or `import_role` modules, or from within tasks using the `include_tasks` or `import_tasks` modules. When you include a role directly within a playbook or a task file, you have the opportunity to pass variables to it, such as the list of packages you want to install.

For example, if you have a role called `other_role` and you want to include the `install_pkgs` role within a task file of `other_role`, you could do it like this:

```yaml
- name: Include install_pkgs to install packages
  include_role:
    name: install_pkgs
  vars:
    packages: "{{ other_role_packages }}"
```

In this case, `other_role_packages` would be a variable defined within `other_role` that contains the list of packages to install. This variable would be passed to the `install_pkgs` role as the `packages` variable.

If you want to keep the `packages` list within a separate defaults file, you can do so, and then reference that list when you include the `install_pkgs` role, as shown above.

If you need further assistance or have more questions about how to structure your Ansible roles and playbooks, feel free to ask.

USER: I want to ensure that packages are installed before the main tasks are run, so it looks like having the call in the meta directory is the correct place. So how do I call install packages from the meta task using the packages variable defined in the defaults pkgs.yml?

ASSISTANT: To ensure that the `install_pkgs` role installs the packages defined in the `gnome_settings` role's `defaults/pkgs.yml` before the main tasks of `gnome_settings` are run, you can define a dependency in the `gnome_settings` role's `meta/main.yml`. This will automatically include the `install_pkgs` role and use the `packages` variable from `gnome_settings` when the role is applied.

Here's how you can define the dependency in `gnome_settings/meta/main.yml`:

```diff
--- vm-setup/roles/gnome_settings/meta/main.yml
+++ vm-setup/roles/gnome_settings/meta/main.yml
@@ ... @@
 dependencies:
 - role: install_pkgs
+  vars:
+    packages: "{{ packages }}"
```

This change assumes that the `packages` variable is defined in the `gnome_settings` role's `defaults/pkgs.yml` and that the `install_pkgs` role is designed to accept a variable named `packages` that contains a list of package names to install.

If the `install_pkgs` role is not currently read-write, please add it to the chat so I can make the necessary edits. If you have any other roles or playbooks that need to be modified to work with this setup, please let me know.